Renaming grapher to graph.
