+++ eccemotus/lib/graph.py
